+++ ProjectLayers/Shared/Interfaces/IUserRepository.cs
@@ -4,10 +4,10 @@ namespace Shared.Interfaces
 {    public interface IUserRepository
     {
         void Add(User user);
-        bool Add(int id, Role role);
-        void Delete(int id);
-        void ChangeRole(int id, Role role);
-        Role CheckUser(int id);
+        bool Add(long id, Role role);
+        void Delete(long id);
+        void ChangeRole(long id, Role role);
+        Role CheckUser(long id);
         List<User> GetAllUsers();
     }
 }
